Move span namer functions to a separate file
